+++ plugins/SqlLogPlugin.php
@@ -15,7 +15,8 @@ namespace AdminNeo;
  */
 class SqlLogPlugin extends Plugin
 {
-	private $filename;
+	/** @var ?string */
+	protected $filename;
 
 	/**
 	 * @param ?string $filename If not set, logs will be written to "$database-log.sql" file.